Removing old files.
