+++ test.js
@@ -1,5 +1,5 @@
 
-var gesd = require('./src/getEquallySpace3').getEquallySpacedData;
+var gesd = require('./src/getEquallySpace').getEquallySpacedData;
 
 var x = [0, 4, 8, 12, 16];
 var y = [2, 5, 3, -1, 7];
@@ -7,13 +7,13 @@ var y = [2, 5, 3, -1, 7];
 var originalOptions = {
     from: -14,
     to: 28,
-    nbPoints: 8
+    numberOfPoints: 8
 };
 
 var options = {
-    from: 0,
-    to: 16,
-    nbPoints: 5
+    from: -14,
+    to: 28,
+    numberOfPoints: 8
 };
 
 var result = gesd(x, y, options);
